refactor: extract session failure extension
Keep ACP orchestration focused by moving the AIR capability contract, failure
mapping, replay helpers, and lifecycle controller into a dedicated module.
This addresses review feedback without changing the negotiated wire behavior.
